Connected posts data to pages, separated posts.js for admin and public pages. fixed url for images.

+++ controllers/post.php
@@ -26,6 +26,10 @@
             if(array_key_exists("id", $_GET)) {
                 getPostById($post_db);
             }
+
+            if(array_key_exists("type", $_GET) && array_key_exists("search_key", $_GET)) {
+                getPostBySearchKey($post_db);
+            }
             break;
         case "DELETE":
             removePost($post_db, $image_db);
@@ -40,6 +44,7 @@
     }
 
     function createPost($post_db, $image_db) {
+        if(!isAdmin()) return "Unauthorized access please login as administrator.";
         $_POST["user_id"] = $_SESSION["id"];
         $_POST["slug_title"] = convertTitleToURL($_POST["title"]);
         $post_id = $post_db->insertPost($_POST);
@@ -47,6 +52,7 @@
     }
 
     function modifyPost($post_db, $image_db) {
+        if(!isAdmin()) return "Unauthorized access please login as administrator.";
         $_POST["slug_title"] = convertTitleToURL($_POST["title"]);
         $post_db->updatePost($_POST);
         if(isset($_FILES["post_image"])) {
@@ -56,12 +62,27 @@
     }
 
     function insertImage($post_id, $image_db) {
+        if(!isAdmin()) return "Unauthorized access please login as administrator.";
         $uploaded_image = $image_db->uploadImage($_FILES["post_image"], FOLDER_ROOT);
         $image_db->insertImage(["post_id" => $post_id, "user_id" => NULL, "image_url" => $uploaded_image]);
     }
 
     function getPostTableData($post_db) {
         $post_res = $post_db->getPostByTypeLimited($_GET);
+        createArrayConvertDataToJSON($post_res);
+    }
+
+    function getPostById($post_db) {
+        $post_res = $post_db->getPostById($_GET);
+        echo json_encode($post_res->fetch_assoc());
+    }
+
+    function getPostBySearchKey($post_db) {
+        $post_res = $post_db->getPostBySearchKey($_GET);
+        createArrayConvertDataToJSON($post_res);
+    }
+
+    function createArrayConvertDataToJSON($post_res) {
         $rows = [];
         if ($post_res->num_rows > 0) {
             while ($row = $post_res->fetch_assoc()) {
@@ -71,12 +92,8 @@
         echo json_encode($rows);
     }
 
-    function getPostById($post_db) {
-        $post_res = $post_db->getPostById($_GET);
-        echo json_encode($post_res->fetch_assoc());
-    }
-
     function removePost($post_db, $image_db) {
+        if(!isAdmin()) return "Unauthorized access please login as administrator.";
         $image_db->deleteImage(["id" => $_GET["image_id"], "image_url" => $_GET["image_url"]]);
         $post_db->deletePost($_GET);
     }
